Create timer prototype
I'm just going to create a timer prototype here. As of the moment, it doesn't do anything other than just prompt the player when a pre-defined time is up.

In the meantime, I'll implement a round a feature where the player must complete a set amount of correct answers to proceed to the next round.